Irreducible noise changes
Modify experiments and the xrd1dim truth functions so as to allow
for mimicing experimental noise. Every time a new point is observed
noise will be introduced randomly to just that point (or not,
depending on the user-set parameters). This is distinct from previous
ways we were doing this, where noise was introduced all at once to
every point in the same way. This meant that even for noisy data,
points close together in X were always close together in Y (in terms
of the differences that the noise produced).